Push valid database reference utilizing MySQL database
Loading folder for RMD file to be able to access data in database based on schema prepared in the initial stage of Project 3. The MySQL database is accessed as a local host and user must have that configured correctly (SET GLOBAL local_infile=1; show global variables like 'local_infile';) on localhost to confirm write privileges are enabled.